chore: :fire: Consolidate training utilities
